refactor: split rendering and generation cleanup roots
Render core, semantic proof suite runtime, semantic surface declarations, synthetic builders, and provider-bundle generation now delegate to narrower helper families instead of mixed-role roots.
